focus when adding a position
so far, when a user clicked on "add position", given that the screen was big enough for desktop mode, then the focus was set to the price of the newly added position. In case of a small screen nothing happened. Changed the behaviour for small screens. Instead of focusing (which would activate the keyboard on mobile phones which is not what we want) we scroll to the price field of the newly added position.